requirements.txt
flake8==3.3.0
requests
pyyaml
email

